[https://github.com/coto/gae-boilerplate/issues/279] Make bp_includes.lib.basehandler share jinja2_factory with bp_includes.lib.error_handler.
